feat: just move/migrate all passcodes theme kotlin code to ths new design system module.
it under active development still. this is not standalone commit
