+++ app/src/main/kotlin/com/jeeldobariya/passcodes/presentation/about_screen/ModernAboutScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.jeeldobariya.passcodes.R
-import com.jeeldobariya.passcodes.ui.ui.theme.PasscodesTheme
+import com.jeeldobariya.passcodes.design_system.theme.PasscodesTheme
 
 @Composable
 fun ModernAboutScreen() {
