+++ app/src/main/kotlin/com/jeeldobariya/passcodes/presentation/main_screen/ClassicalMainScreen.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.unit.dp
 import com.jeeldobariya.passcodes.R
 import com.jeeldobariya.passcodes.core.navigation.Route
 import com.jeeldobariya.passcodes.design_system.DESIGN_SYSTEM_NAME
-import com.jeeldobariya.passcodes.ui.ui.theme.PasscodesTheme
+import com.jeeldobariya.passcodes.design_system.theme.PasscodesTheme
 
 @Composable
 fun ClassicalMainScreen(navigateTo: (Route) -> Unit) {
