+++ app/src/main/kotlin/com/jeeldobariya/passcodes/presentation/main_screen/ModernMainScreen.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.jeeldobariya.passcodes.R
 import com.jeeldobariya.passcodes.core.navigation.Route
-import com.jeeldobariya.passcodes.ui.ui.theme.PasscodesTheme
+import com.jeeldobariya.passcodes.design_system.theme.PasscodesTheme
 
 
 @Composable
