+++ password_manager/src/main/kotlin/com/jeeldobariya/passcodes/password_manager/presentation/password_manager/ModernPasswordManagerScreen.kt
@@ -32,6 +32,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.jeeldobariya.passcodes.core.navigation.Route
+import com.jeeldobariya.passcodes.design_system.theme.PasscodesTheme
 import kotlinx.coroutines.launch
 import org.koin.compose.viewmodel.koinViewModel
 
@@ -113,8 +114,10 @@ private fun ModernPasswordManagerScreenContent(
 @Preview
 @Composable
 private fun ModernPasswordManagerScreenPreview() {
-    ModernPasswordManagerScreenContent(
-        state = PasswordManagerState(),
-        navigateTo = {}
-    )
+    PasscodesTheme {
+        ModernPasswordManagerScreenContent(
+            state = PasswordManagerState(),
+            navigateTo = {}
+        )
+    }
 }
