+++ password_manager/src/main/kotlin/com/jeeldobariya/passcodes/password_manager/presentation/save_password/ModernSavePasswordScreen.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.jeeldobariya.passcodes.design_system.theme.PasscodesTheme
 import kotlinx.coroutines.launch
 import org.koin.compose.viewmodel.koinViewModel
 
@@ -112,5 +113,7 @@ private fun ModernSavePasswordScreenContent(
 @Preview
 @Composable
 private fun ModernSavePasswordScreenPreview() {
-    ModernSavePasswordScreenContent(state = SavePasswordState(), onAction = {})
+    PasscodesTheme {
+        ModernSavePasswordScreenContent(state = SavePasswordState(), onAction = {})
+    }
 }
