+++ password_manager/src/main/kotlin/com/jeeldobariya/passcodes/password_manager/presentation/update_password/ClassicalUpdatePasswordScreen.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.jeeldobariya.passcodes.design_system.theme.PasscodesTheme
 import kotlinx.coroutines.launch
 import org.koin.compose.viewmodel.koinViewModel
 
@@ -117,5 +118,7 @@ private fun ClassicalUpdatePasswordScreenContent(
 @Preview
 @Composable
 private fun ClassicalUpdatePasswordScreenPreview() {
-    ClassicalUpdatePasswordScreenContent(state = UpdatePasswordState(), onAction = {})
+    PasscodesTheme {
+        ClassicalUpdatePasswordScreenContent(state = UpdatePasswordState(), onAction = {})
+    }
 }
